+++ Funcky/Extensions/QueueExtensions.cs
@@ -1,0 +1,52 @@
+using System.Collections.Concurrent;
+
+#if !QUEUE_TRY_OVERLOADS
+using Funcky.Internal;
+#endif
+
+namespace Funcky.Extensions
+{
+    public static partial class QueueExtensions
+    {
+#if QUEUE_TRY_OVERLOADS
+        [Pure]
+        public static Option<TItem> DequeueOrNone<TItem>(this Queue<TItem> queue)
+            where TItem : notnull
+            => queue.TryDequeue(out var result)
+                ? result
+                : Option<TItem>.None();
+
+        [Pure]
+        public static Option<TItem> PeekOrNone<TItem>(this Queue<TItem> queue)
+            where TItem : notnull
+            => queue.TryPeek(out var result)
+                ? result
+                : Option<TItem>.None();
+#else
+        [Pure]
+        public static Option<TItem> DequeueOrNone<TItem>(this Queue<TItem> queue)
+            where TItem : notnull
+            => HandleException<InvalidOperationException>.ToNone(queue.Dequeue);
+
+        [Pure]
+        public static Option<TItem> PeekOrNone<TItem>(this Queue<TItem> queue)
+            where TItem : notnull
+            => HandleException<InvalidOperationException>.ToNone(queue.Dequeue);
+
+#endif
+
+        [Pure]
+        public static Option<TItem> DequeueOrNone<TItem>(this ConcurrentQueue<TItem> concurrentQueue)
+            where TItem : notnull
+            => concurrentQueue.TryDequeue(out var result)
+                ? result
+                : Option<TItem>.None();
+
+        [Pure]
+        public static Option<TItem> PeekOrNone<TItem>(this ConcurrentQueue<TItem> concurrentQueue)
+            where TItem : notnull
+            => concurrentQueue.TryPeek(out var result)
+                ? result
+                : Option<TItem>.None();
+    }
+}