+++ Funcky/Internal/HandleException.cs
@@ -1,0 +1,19 @@
+namespace Funcky.Internal
+{
+    internal class HandleException<TException>
+        where TException : Exception
+    {
+        public static Option<TResult> ToNone<TResult>(Func<TResult> func)
+            where TResult : notnull
+        {
+            try
+            {
+                return func();
+            }
+            catch (TException)
+            {
+                return Option<TResult>.None();
+            }
+        }
+    }
+}